5700 - update variable box to use static area type title
