+++ helpers/helpers_test.go
@@ -83,3 +83,23 @@ func TestGenerateSharingLink(t *testing.T) {
 		So(GenerateSharingLink("email", url, title), ShouldResemble, fmt.Sprintf("mailto:?subject=%s&body=%s\n%s", title, title, url))
 	})
 }
+
+func TestIsBoolPtr(t *testing.T) {
+	Convey("When the value is nil", t, func() {
+		Convey("Then the returned value is false", func() {
+			So(IsBoolPtr(nil), ShouldBeFalse)
+		})
+	})
+	Convey("When the value is a false pointer", t, func() {
+		ptr := *new(bool)
+		Convey("Then the returned value is false", func() {
+			So(IsBoolPtr(&ptr), ShouldBeFalse)
+		})
+	})
+	Convey("When the value is a true pointer", t, func() {
+		ptr := bool(true)
+		Convey("Then the returned value is true", func() {
+			So(IsBoolPtr(&ptr), ShouldBeTrue)
+		})
+	})
+}
